backend/ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
pymongo==4.12.0
numpy>=1.26.0,<2.0.0
pdfplumber==0.11.6
rank-bm25==0.2.2
faiss-cpu==1.11.0
google-generativeai==0.8.4
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-community==0.3.5
langgraph==0.2.45
